license__Seamless/requirements.txt
pyTelegramBotAPI
python-dotenv
requests
flask
